Make Duke follow OOP
To follow OOP, we make the Duke bot an Object, which can be instantiated multiple times. This allows us to create instances of the Duke bot elsewhere, even from other applications.

To further improve code clarity, I moved the entry point of the program to the Main class, along with the Main method. The Duke class has further been renamed to HelpBot to allow anyone to name the bot however they like.

In my case, I have specified the logo in the main class as a constant, that I now pass through to the HelpBot class as the bot's name.

Finally, javadocs have been added to the classes for documentation purposes, and the test batch file has been edited to run the new main class.