Add ability to mark tasks as done
Right now, the tasks are kept track of in a List of Strings, and the number of commands has increased to 3, with new and more complex commands being added.

To allow us to easily keep track of tasks and their status, it is easier to create a new Task object to keep track of this information, instead of storing the tasks in a List. We therefore add Task.java to create these objects, instantiating them with a name and completion value of false.

Since the number of commands is starting to grow as well, an enum containing all the types of commands is now a better way of keeping the user command input method easy to understand. To use this enum, we add getCommand to HelpBot.java to parse the text inputs for command variables. In future, we can easily edit this method if we ever want to add more command parsing to our bot, and we can easily edit the enum to add more commands.

With the new command requiring us to get an integer from the second word of the command, and future commands possibly needing this too, we add a new class Methods.java to contain commonly used methods across multiple classes. The first method we add here is isInt(String message), which checks if a given String is an int.