Add multiple types of tasks
Currently, we have one type of task, which may not be enough for users who want to set deadlines for their tasks, or plan their time better.

To make sure that the code is readable, we create three classes, one for each of the new task types. They all extend the main Task.java class, so we can easily call methods from the main class to access the other classes.

Since users now have access to tasks that have dates as well, users may change their mind on when they want these tasks to be fulfilled. To allow this, we add the ability to update the deadline and date for the deadlines and events respectively.

To keep track of tasks to edit, we use our List of all the tasks, which we can use to easily reference tasks in the correct index to update. An abstract method updateDate in Task.java is added to allow its children to inherit the method. The Todo.java class throws an IllegalStateException when this method is called since it does not have the date field required to be updated.

We add the method getTaskAmount to return us the number of total tasks when we add a new task. We also add the method onEnableMessage to handle sending the welcome message to new users, now with instructions on how to use the bot, of course with the same passive aggressive spin.