Amend testing files
Presently, javac is used to compile the files and test them. This was easier when the app was not in multiple packages, however with the splitting of classes into different packages, it has become easier to build an artifact and use the java -jar command to run that instead.

To do this, I build the project artifact in the IDE, which gives me the updated ip.jar. This allows me to run the script test.bat on it, which takes in the input and creates the file ACTUAL.TXT. Finally, the FC results are saved to results.txt to allow us to view it afterwards.