Add error handling
Currently, error handling uses some of Java's own errors, as well as methods situated in the Methods.java class. To make errors more useful and informative to users, we should create new errors specific to the bot.

To do this, we create the class HelpBotException.java, which is the overarching class for all of the errors the bot may face. We divide this into two cases for now, the InvalidTaskType and InvalidTask exceptions. The former is called if the user tries to perform an action on the wrong task type, while the latter is called if the user tries to perform an action on a task that does not exist.

Following the changes, we remove Methods.java as error handling has been moved elsewhere. Update EXPECTED.TXT to reflect new messages related to errors.