Improve commands and errors
Presently, we do not have a command for the default case of adding tasks to the bot. This could result in confusing cases where users type something wrong into the bot and the typo gets added as a task.

To fix this, we add a new command "todo" to handle the case where users want to add a todo item to their task list. The default case for user messages to the bot is now an error telling the user that the entered command is invalid.

We also add other error classes to help differentiate between the new errors users may face. We add a new HelpBotIllegalArgumentException as a general error that refers to commands with wrong arguments, with InvalidTaskException being a more specific form relating to the task specified by a command being invalid.

Update onEnableMessage to reflect new command and add documentation to the method. Update test cases to reflect the changes. Update .gitignore to stop pushing results.txt to remote.