Add task deleting functionality
With the ability to mark tasks as completed a few updates ago, users can now keep track of their tasks. However, this could result in clutter if the completed tasks remain on the list of tasks.

Add deleteTask method to Task.java to allow users to delete tasks, as well as handle messaging users for a reply. Move getTaskAmount method to Task.java since the tasks are stored there.

Rename IllegalCommandException to InvalidCommandException to more accurately reflect the nature of the exception. Rewrote reply message for InvalidTaskException for more clarity when debugging.

Updated test files to match updates.