Add Level 2 Features to the chatbot
It would be good if we can store the user entered
inputs, and the user could view those any
time he would like to do so.

Add an List<String> to store the various
user inputs. Register some keywords
including "bye" and "list" to terminate
the session with the chatbot and to view
the items present in the list respectively.

List<String> is used as we can store
almost infinite inputs versus how arrays
would restrict to a certain number that has
to be determined at the start of the program.