Add Level 3 features to the Chatbot
To make the chatbot more functional, and to help
you through your daily chores, this chatbot can now
keep a track of various tasks you have.

A new class called Task is added to help track the
task status(whether done or not). Also, the ability to
mark tasks as done is added.