Duke class: run processor in a while loop
Duke class processes only one command and exits.

The program shouldn't exit after processing one command,
it should continue to run unless the 'bye' command is received.

Adding a while loop in `main` allows the program to keep running.
The return type of method `process` in class `Processor` is
changed from void to boolean, the method returns false only if
the command is 'bye'.

Changing the return type from void to boolean enables the while
loop to stop when 'bye' command is received.