Add Level 4 features to the Chatbot
There are various types of tasks including but
not limited to events, deadlines and todos.
The chatbot should be able to serve the end user
better and be more functional.

Extracting the common behaviour into a super class,
called "Task" allows to use polymorphism when dealing
with various types of tasks.