Extract scanner from Event.java
Each event contains a scanner to read command.

This shouldn't be the case because an event should be triggered
after a command is received.

Scanner is removed from the Event class. A method called `readCommand`
is introduced in the Duke class, which will read the command and
decide the Duke program's next action.

Moving the reading command process into the Duke class enables
the program to decide which event to trigger based on the command.