Duke class: extract processCommand method
The steps of calling the process method of a processor and print
the result are repeated for each command.

These lines of code are repeated, which should be avoided.

In the Duke class, a method called processCommand is defined to
contain the repeated lines of code. Then the common lines in the
readCommand method are replaced by calling processCommand.

This makes it easier to debug and change the behavior later.