Add Level 5 features to the Chatbot
The Chatbot should not crash at any instance if
an invalid/unformatted input is entered by the user.
It should instead inform the user to use the right
format and perhaps show an error message as well.

Custom exceptions are made to take care of various
scenarios where a user might enter a wrong input.

Also updated input and EXPECTED text files for
testing purpose.