Duke class: handle exceptions
Invalid commands such as 'todo' and 'blah' are allowed, and these
will be treated as a todo task.

Allowing invalid commands may lead to problems such as recording a
typo as a todo task.

The following exceptions are handled using DukeException class:
- Invalid commands (commands other than bye, list, done, todo,
  deadline, event)
- Invalid task number in a done command
- Empty description in todo, deadline, event commands
- Empty time in deadline and event commands
- multiple time slots in deadline and event commands

By creating a new class, the error message can be changed into
the same format as other messages. Handling errors also provide
the users with a guideline on how to use the program.