Feat: Add task types
Each task type is a class which inherits from classes. These are todo,
deadline, and event; deadline has a by property which is when it is due
and event has an at property which is when the event is held.

These are created by starting the input with the keyword.

Also refactored the way the input was parsed.

Solves level-4 and A-inheritance