Add basic saving features
The bot does not have any way to save the given tasks. This results in the user having
to repeatedly enter his tasks.

Add saving mechanisms. Create the relevant folder and file (createDirectory()). Write the
tasks into the .txt file (retrieveTasks()) and read from it when the bot start (if possible).

It is being done this way as having multiple methods handle the writing/reading aspect
increasing debugging ability and readability.