Add more OOP
Right now, we do not have our commands, task list and command parser in separate classes. This means that as we add more commands, the method count in the HelpBot class will keep increasing, making it increasingly hard to keep track of.

To fix this, we create new Command objects for each of the individual commands, which we can run and handle errors easily for. The new Ui class handles text replies and formatting. The TaskList class is created to do operations on tasks and keep track of the list of tasks. Finally, the Parser class centralises the parsing of commands.

This centralises error handling to HelpBot class to allow us to see at a glance what goes wrong. Tracing errors to individual commands is also easier. New commands are easily added to the Parser and new children of the Command class can be created for commands added in future.

Update test files to match updates.