Format and abstract response, fix bugs
Some responses are quite long and cluttered up the Petal class, reducing readability.
The strings returned by some methods (Task toString()) and printList() are not in the
format required for saving. There are bugs in the displayed list after accessing saved tasks
because of the wrong format being used to save the tasks.

Add abstract methods strForSaving() that returns a string representation for the task in the
proper format (e.g. T|X|go for a run). Add formatForSaving() that prints using this method.

Used enums as it always for constants and used abstract methods to take advantage
of polymorphism.