Add Javadocs
Right now, after adding more classes to follow OOP, we have not updated our Javadocs in the new classes to match what we have done before. The convention of Javadocs was also not correct before.

Add Javadocs to all methods and classses, and comments to all variables to tell myself and other developers what they mean. Methods and classes that are inherited from others or extend others are not included. Also make slight changes to bot's responses for personality reasons.

This allows us to pick up right where we left off if we do not work on the project for a few days. Other developers can easily use or hook into our project to further build on it.

Update test files to match updates.