Add Storage.java
Right now, we only accept user input from System.in. Although users can specify files to move files in, we could add the functionality to the bot to automatically load commands and save results to files.

Add Storage.java to handle IO to and from files. Because of this, we rewrite some classes to allow for the new file saving functionality, mainly in Ui, where we save the replies from our bot. We still keep the System.in input format if a file is not provided for input.

This allows users to have an easier time using our bot, since they can just copy the file over and run the jar file instead of using the command line to do so. Keeping System.in as an alternative form of input allows users of the bot to still use it without having to put all their commands in a file.

Update test files to match updates.