Add time feature and add test cases
The code can only handle time in a string form, and is unable to do anything more with it.

Store time as a LocalDate and LocalTime object. Add additional test cases to test for bugs
in time input.

Storing time as a LocalDate and LocalTime object gives much more flexibility. allowing
Petal to display time in a specific format, and in the future, track events based on the data