Save the tasks in the hard disk automatically
The tasks are only temporarily used during one execution of the program.

The program needs to be more reusable so that it can actually keep
track of all the tasks that have been made so far in the lifespan of
the program.

A storage for the list of tasks can be implemented to make the program
more reusable.