Retrieve tasks on given date (deadline)
Petal can only associate a deadline with a given task, but cannot return an output
with all the deadline on a given day.

Add HashMap. Use static method that takes in a date and returns string
representation of value stored at date.

The HashMap allows for the use of key-value pairs that will store the corresponding dead-
lines at that date. Public static method will allow the method to be called from the Petal
class and return the string representation.