Add File Saving/Loading
Right now, the bot does not save the task list for users after they have used the bot. THis means that users lose their todo list everytime they close the bot.

We improved this by adding file writing and reading functionality for our list of tasks to the Storage class. This updates the file in real time as the user types commands into the bot.

Doing this allows users to keep track of their tasks across different sessions of starting our bot. Updating the file in real time may be more resource intensive but it makes sure no data is lost in the event of a crash or accidental closure of the bot.