Retrieve events and deadlines on given date
The code can only retrieve deadlines on a specified date, but not events.

Add similar methods to event class to parse given input into date, start time, and end time.
Use hashmap to keep track of events on that date.

Similar reasons to the deadline commit. Refers to the prefer commit for more information.