Update Deadline and Event objects with DateTime.
Deadline and Event object previously took time as a String.

Code has now been updated such that time attribute of Event and
deadline attribute of Deadline are now LocalDateTime objects.

Duke now takes event and deadline inputs in the following manner:
* Event - "event *description* /at *DD/MM/YYYY* *24H-Time*
* Deadline - "Deadline *description* /by *DD/MM/YYYY* *24H-Time*
If time is not included in the input, a default time of 0000 is used.

Also updated input.txt and EXPECTED.TXT in text-ui-test accordingly.