Create hard disk file to save task list
Adding the ability to save tasks between separate uses of the program
improves usability, rather than have to start a new task list from
scratch every time.

We start by creating a file 'duke.txt' which acts as the hard disk
in a new directory 'data', and allow Duke to check if the hard disk
exists, create it if not, and access it.

Let's
* create the folder 'data' in the program root folder and the file
'duke.txt' inside 'data' to act as a hard disk and store the task list
* add functionality to Duke to check if 'data' and 'duke.txt' already
exist and create them if they don't already exist
* add exception handling in case the program is unable to create 'data'
and 'duke.txt'