Add Date/Time support
Right now, the bot mere takes in date and time as strings. While this saves on a lot of backend error checking and parsing, it also means the information could be less useful for the user.

We improve this by making sure users input actual date and times in AddCommand. This means users now have to provide actual specific timings for their tasks. We make a new DateTimePair class to keep track of this within each Event and Deadline object. A new error is created to handle errors related to parsing date and time.

Doing this allows users to have a more consistent experience, with all date and time print outs being in the same format. We are also able to fetch tasks by date, allowing users to see at a glance what they have to do on a certain day.

Minor bug fixes have been made to deletion of task and parsing of various commands.