Create methods for interacting with hard disk
Duke needs methods to interact with the hard disk whenever a change
is made to the task list, allowing the hard disk to keep track of them.

Let's create the methods addEntry, setEntryDone, and deleteEntry to
add, delete and set tasks to done on the hard disk.

After this these methods will be implemented into Duke's main function.

There is also a need to abstract out the process of checking duke.txt
upon starting the program, and adding the functionality of loading in
the tasks saved to the hard disk.