Implement hard disk interaction methods in Duke
Implement the methods we added in the previous commit in the Duke
function so any changes to the task list in Duke are reflected in the
hard disk as well.

Also, create a custom string format for the tasks added to the hard
disk for readability.

Finally, add exception cases to DukeException to deal with new potential
exceptions.

Let's
* implement addEntry, setEntryDone and deleteEntry in the add, done
and delete logic of the Duke.main() function
* create a new string format for the different tasks for adding to the
hard disk
* implement new exception cases to deal with potential exceptions
arising from the new methods