Fix bugs and move scanner
After abstraction, the bye command no longer works.

Add public method in Petal to terminate bot. Move scanner to Ui class as part of
OOP.

The public method allows the Ui class to toggle the boolean Bye, turning it off and moving
the scanner to Ui allows for better abstraction.