Add storage for task lists
After program exits, list of tasks is not saved.

By storing the list of tasks in a text file locally on your computer.
Even after the program exits, the bot can recreate the task list from
the text file, when users interact with the bot again.

This is more sensible, as users will not have their task lists wiped,
and will not have to recreate the list again.