Add dates and times to chatbot
Chatbot saves the timing of certain tasks as strings.

The timings are now saved as LocalDate objects.

This allows for more flexibility and ease when we implement future
commands such as showing tasks that occur within a given time period, as
the class LocalDate has a lot of its own built-in fuctionalities.