Implement the TaskList class
Implement TaskList to handle any and all functionality to do with
altering the task list. In addition, TaskList will also handle the
changes made to the storage of tasks by Storage.

Let's extract any functionality to do with the task list (methods to
add, delete, make changes to the task list, as well as UI console output
methods upon task list changes) for the TaskList class to handle.