Add more OOPs and organize into packages
Add Parser, Command and subclasses of Commands. Package catobot
and command.

Organization into packages and addition of OOPs are done in one
commit because they are closely linked to another. Without
organizing Catbot, Storage and Ui into packages, it would be
hard to create the interface Executable and it would be hard to
maintain a desirable OOP.