Implement the Parser class
The Parser class takes over the Duke class's job of taking in an
instruction from the user, interpreting it and performing the relevant
action after.

It will either return a Command (to be implemented) for the Duke class
to perform, or throw an exception to the Duke class to handle.

Let's implement the Parser class to parse user input and return a
Command to be executed.