Extract different commands from the Command class
For greater clarity, the Command class can be made into an abstract
class, and create more specific Command classes (eg. AddCommand) to
handle each type of command.

Let's
* create new classes: AddCommand, ByeCommand, DeleteCommand,
DoneCommand, FindCommand and ListCommand, that extend Command, which is
made abstract, and extract the functionality from the original class
to the child classes
* modify Parser to use the new classes