Reorganise test files
Currently, the test files are all in the same directory with a flat
structure.

This is disorganised and hard to read.

This commit will organise test files into subdirectories.

Reorganisation makes it easier to navigate the test files.

The test should be run from the project root directory.
The file to run is `./text-ui-test/duke-test.sh`