Extract Ui, Storage, Parser, TaskList and Message
Duke contains too many methods. Some of the methods are just for
interacting with the user; some are just loading and saving tasks to
file; some are just dealing with user inputs; some are involved with
manipulating the list of tasks; the rest is just printing out responses
made by the chatbot.

This causes the code to be hard to follow along.

Extracting the methods into their respective classes delegates
responsibilities to the classes and makes it easier to refactor certain
behaviour of Duke by just refactoring only the class that handles this
behaviour.