Add A-Packages to the Chatbot
The files were all organised in one
folder. Searching for files was becoming
complex as the number of files kept on
increasing.

Add packages. Organize the files into
exception, task, and retriever packages
respectively.

To avoid naming conflicts and
to make it easier to search for various files,
packaging has been implemented.