Make the code more OOP
Refactor the code to extract out closely related code as classes.
Ui: deals with interactions with the user.
Storage: deals with loading tasks and saving tasks in the file.
Parser: deals with making sense of the user command.
TaskList: contains the task list and has operations to add/delete tasks.