Standardising name of program
Change all instances of the original intended name of the program, Duck,
to Duke, which I started using after a while for some reason.

Let's change all 'Duck' to 'Duke'.

If I have the time to, I might change the program name back to Duck at
some point, but that will take time.