Add A-JUnit: Add JUnit Tests.
Duchess uses I/O redirection to test.

This is undesirable as the project grows and the texts become
more complex and difficult to compare with the expected output.

Let's implement JUnit Tests to test the behaviour of the code.
Let's add Test classes for the different classes within Duchess and
package them into respective packages.