Add timed status to tasks
Currently, tasks with a time (e.g. Event and Deadline) are not differentiated in the code from tasks without a time (e.g. ToDo).

This makes it difficult to add a time formatting function to Duke, as it is difficult to determine which tasks need an extra formatter to print.

By adding new timed property, it becomes easier to print tasks with a time formatter