Refactor code to be more OOP
Added a Ui class that deals with interactions with the user.
Added a Storage class that deals with loading tasks from the file and
saving tasks in the file.
Added a Parser class that deals with making sense of the user command.
Added a TaskList class that contains the task list.