Implement find functionality
Duke did not have a search functionality which made it hard for users to find previously made tasks.

The introduction of this command will help improve user friendliness of Duke.

Let's implement find functionality as a Command type and update the UI and Tasklistss classes to be able to accommodate this functionality.