Add varargs feature
Users can now mark multiple tasks as completed or delete multiple tasks in one command by giving a list of index.

Previously if a user wanted to mark task 1, 2, 3 and 4 as done, the user had to enter 4 separate commands.
Now, the user can just enter done 1 2 3 4 in order to mark the 4 tasks as completed.