Tasklist: add duplicate checks
Tasklist allows identical tasks (e.g. todos with the same name) to be
added.

This worsens the user experience as users need to keep track of
duplicate tasks by themselves.

Encapsulating the list of tasks in its own class allows
additional behavior to be added that is not provided by a regular
ArrayList. One of this is checking of duplicates when inserting into
the list.

Let's create a class called TaskList that encapsulates the
aforementioned behavior.