Assertions: add assertions using assert feature
Current code will not terminate the code execution if a possible bug is discovered. This could be unsafe.

Assertions are used to define assumptions about the program state so that the runtime can verify them.

Let's
* do some refactoring to abstract variables as suitable checkpoints for assertions
* add assertions in the classes Duke, Parser, Storage and TaskList.

Refer to this resource on the importance of assertions and how to include them:
https://nus-cs2103-ay2122s1.github.io/website/schedule/week5/topics.html#w5-6-assertions