Storage class: add an assert feature
Method .load() lacks an assert feature

This method makes a critical assumption that
every line in the memory text file is formatted
correctly, without being an empty line.

Assert feature ensures that this assumption
stays valid.

This helps in a debugging process, wherein
the text file was not saved properly by
.store() method.