Implement GUI for Level-10
The text-based UI is lacking in terms of user interaction and Java-FX
seems to be able to give a better user experience.

Implement a text window and text bar to display the user and bot's
conversation.

This format of displaying text is similar to common mobile text
applications which is easy for users to get used to.