printList and showSearchResults: add streams
Current printList and showSearchResults are cumbersome and involve verbose for loops over the ArrayList of Tasks.

Using streams to map each index in the list to the correct String format makes for cleaner code and increases code quality.

Streams are also used to show understanding of new Java 8 features.

Let's
* Rewrite the existing printList and showSearchResults in the Storage and Ui classes respectively
* Test the implementation via GUI