Rename methods in Ui.java
New methods created for GUI have different naming convention that starts
with "show" and old methods using CLI starts with "print".

The inconsistencies potentially reduces readability and ease of
understanding for future managers of the code. Hence a standardization
is required.

Rename all "print" prefix methods to "show" prefix with minor
adjustment to some method names.

Using "show" is more suitable to describe the actions of duke because
Duke now runs on GUI instead of CLI.