Implement Assertions
Checks are not being done to ensure that certain String inputs
are sanitized.

There are methods in the app that receive Strings as input.
Assertions help to ensure these Strings are formatted properly
(i.e. lowercase).

Write assertions to help to guide the developer on the
expected input/outputs of the code.

Assertions are written for methods involving Strings, since much of the
application involves user-supplied text for interaction.