Create new Github Action Workflows.
Checkstyle and Tests are not automatically checked on PRs and pushes.
Fat JARs have to be manually built and uploaded to each version release.

Repetitive manual actions like these waste developer's time.

Create Github action workflows to automatically have Gradle run the
`check` task to verify PRs and pushes to master maintain repo quality.
Then the `shadowJar` task is used to create the fat JAR and then upload
to the `latest` release.

Gradle and shadowJar already has well implemented ways to test and
create fat JARs and using Github's action process automates this.

Note that for shadowJar, it runs on version releases, which are
identified as a tag named 'v*' (e.g. v1.0).