Add C-Sort functionality.
Tasks are presented in the order they are added.

User will likely want to see urgent deadlines and events before other
tasks.

Add an abstract function to `Task` to obtain a numeric value of urgency.
Then provide a comparator in `Task` for use in `Parser` to sort by
urgency before presenting users with a list.

To show this view, Duke must reorder the tasks in terms of the date
associated with it. Therefore, the different Task subclasses must
provide a numeric value to compare dates against each other.
Converting them to Unix timestamp results in an easily handled number.
This also lets Todo tasks (without any dates) to occupy the bottom with
`Long.MAX` as it has no date (i.e. no urgency).

In the future, any comparators can involve other attributes like
priority in its calculation to create a final comparison.