Add assertions to validate assumptions in code
Assumptions are that users will not pass null as a String argument to Task-related
constructors or to parseSaveFormat method.
Similarly, assumptions are also made that users will not pass empty string or null into DialogBox constructor
and that they will also not pass a null as Image argument into DialogBox constructor.
Assertions need to be made to ensure that later users do not misuse these methods/constructors.

Assertions are added to corresponding constructors/methods to validate the aforementioned assumptions.