Add GUI
Right now, we use the CLI for task entry, which is not aesthetically appealing to users. It is also hard for users to tell which comamnd is from the Bot.

To fix this, we create a new Window for all our messages and replies, as well as input and sending. The user and bot also have avatars now, which correspond to who sent the message.

This allows users to easily keep track of tasks, as well as enjoy chatting with the bot more.