Add Varargs
Right now, deletion of tasks is a tedious process, especially with the bug I introduced last update, which adds new tasks twice.

To fix this, I fix the bug first, then add the ability to parse multiple task numbers through to the delete command via the use of varargs.

This allows users to delete many tasks at once, and saves on time. The bot also has some new messages to go with this change.