Add Command abstractions
Parser currently does all of the work to interact with the logger and
tasklist.

In the long term, if there were new commands, the switch statement would
have to be very long. If there were specific changes to commands, it
would be difficult to edit.

Abstraction of the commands into different classes can allow for more
manageable code in the long term.