Add Assertions to several methods
Use assert feature to document important assumptions that should hold at various points in the code to reduce the occurrences of bugs in the program.

Let's put Assert() methods throughout the code in appropriate places.

This is a quick way to ensure that errors made by us are discovered quickly. For more in-depth testing, JUnit tests will be used.