Implement Friendlier Syntax extension
Commands are strictly matched and the parser uses many if/else
statements to check which command to invoke.

Refactor the `Parser.java` class to be more flexible with various commands, as
well as give the `CommandType` enum room to store aliases for commands.

Implementing such flexibility allows for easier extensions of more
command types and command aliases in the future.

The idea for extending the enum came from
[this StackOverflow post](https://stackoverflow.com/questions/4197988/java-enum-valueof-with-multiple-values/4198066#4198066).