Improve code quality
Making the code less nested improves the code quality as code that is deeply nested is not very optimal.

The code is also shortened (but still has the same functionality) to improve readability and also improve code quality.
This is done by using DateTimeFormatter to shorten and make the getTimeDate method more efficient.

Some variable names are also changed to make it easier to understand the purpose of the variables.