Assert that Task to be added is not null
TaskManager::addTask assumes that the Task argument is a
valid Task and adds it directly to the taskList.

This makes null a valid argument and may cause unexpected exceptions
when we operate on it.

Let's add an assertion at the start of TaskManager::addTask that the
Task to be added cannot be null.