Use streams to format string representation of message
The toString() method combines an array of strings into one string using a for loop with an index to iterate.

It is easier to understand the function by converting the array to a stream then use a for loop. This is because the index is only used for iteration and nothing else.

Let's,
* convert the array of strings to a stream
* iterate through the stream using forEach