Extract formatting of tasks as list into function
The overloaded methods of TaskManager::list filter the tasks (if
applicable), get the string representations of the individual tasks in
an array of strings and then join the strings to form the returned
string.

This common behaviour causes code duplication.

Extracting this behaviour into a function allows us to reduce code
duplication by calling this function from each list method.

Let's create a method formatTasksAsList to address the code duplication.