Add bye command
Users can only exit the app by clicking on the cross button.

It is not efficient for users to do so since the app only requires the
use of keyboard.

Let's add a bye command that will allow users to exit the app without
having to switch to a mouse.

It is more efficient for users to continue using a keyboard to exit
instead of switching to a mouse.