Add assertions
There are several methods that work based on certain assumptions which
are thought to be checked by their calling methods.

The programmer might leave out certain cases during the checking
process, leading to bugs when the method is called.

Let's add assertions to ensure that these critical assumptions hold true
at the point of method call.

Using assertions help the programmer to identify where mistakes might be
made, and also terminates the app if such mistakes are made to prevent
buggy method calls.