Add Assert statements to certain points in code
No asserts in current code.

It is good to check assumptions about program state so runtime can verify them.

Add some assertions in crucial parts of the program.

The program can verify it is running properly during runtime and terminate in the case of assertion failure to indicate possible bug.