Add assertions.
There are comments signifying certain invariants, as well as default
cases in switch blocks that have an empty body.

Comments do not functionally stop the code if the commented invariant is
violated. Empty default cases in switch blocks also work under
assumptions that are not watched out for by the compiler.

Let's substitute comments with assertions, as well as add assertions
into the empty default case bodies.

Using assertions is preferable over comments to result in the program
telling us if invariants were to be violated.