Add assertions to document important assumptions
There are a few methods that have assumptions in their body, either
through switch statements, if-else statements or that another method
returns a value that is supposed to be valid.

Having assertions in these methods would document these important
assumptions that has to hold at various points in the code.

Let's insert assertions for,
* internal invariants
* control-flow invariants
* preconditions
* postconditions
* class invariants