Remove misuse of assert statements and add new assert statements for DialogBox flip method
Assert statements were used to validate that arguments passed into public methods were not null,
when it is more appropriate to let it raise runtime exceptions like NullPointerException that describe
the error better.

Remove misused assert statements and add assert statements for flip method of DialogBox.

Flip method of DialogBox assumes that the HBox is already aligned to the top right of its container.