Refactor code to allow for exception to take in no arguments
Previously, several exceptions took in an empty string even though the output message did not depend on the string.

To remedy this, the constructors of several exception functions were changed so that it accepts no arguments, if the arguments are not used in the message the exception generations.

This improves the readability of the code.