Use Assertions
Assertions are not used in this project.

A project not using assertions may be vulnerable to errors caused by
assumptions made by the software engineer.

Let's add assertions whenever an assumption of invariants or conditions
is made.

In this case, assertions are more appropriate then exception handling
as we are targeting mistakes made by programmer rather than the user.