Add CI
Currently, there is no CI being done.

This needs to change so that everytime one does a PR or git push, the new changes do not break the code.

As such, a Github workflow is added so that it builds, runs the unit tests, and runs the text ui tests every time a Github action is take.

This is to minimise bugs.