Add assertions
The code does not have any assertions, and this may cause issues later now as
the states of certain attributes and objects is assumed (not verified) which may cause
unwanted side effects.

Add assertions liberally throughout the code (especially to assert if a attribute is not null,
for example in the case of 'assert mainWindow != null' in the Ui class.

It is being done this way in accordance with the documentation, which states that asserts
should test for certain states in the program.