Fix bug when no tasks are found
When the user enters in a search term, and no tasks are returned, Tiger previously responsed with "You currently have no tasks".

This is wrong and confusing as it implies the user has no tasks, rather than that no tasks match the find description.

This new commit fixes that by first checking if the taskList returned when the user searches for tasks is empty. If it is empty, Tiger responds instead with "I couldn't find any tasks."