Refactor DukeGui to improve readability
Currently, the handleUserInput method in DukeGui is easily confused with
the handleUserInput method in Duke (which extends DukeGui). Neither
overrides the other and they both accomplish different functionalities.

The handleUserInput method in DukeGui needs to be renamed and refactored
to show that it simply echoes the user input, and does not carry out any
additional computation about it.

Let's rename handleUserInput in DukeGui to 'echo', and remove the
dukeText variable to show that the same text is used for input and
output.

This renaming and refactoring clearly shows the difference between the
logic in DukeGui and Duke. Confusion is reduced since Duke extends
DukeGui and we clearly show that there should be no method overriding in
handleUserInput.